dockerbundle: support Gitlab CI DIND use case
This adds support for DIND implementation in Gitlab. In Gitlab CI,
access to Docker is possible by using DIND instance. This DIND instance
is setup by the Docker Runner when specified as a service in the Gitlab
CI yaml file. In a first attempt to implement Gitlab CI support I tried
to make use of this DIND instance to create the storage tarball. However,
the storage backend of that DIND container is a Docker volume on the
host Docker instnace and it is not possible to access that Docker volume
from within the Gitlab CI job container. Unfortunately one cannot bind
mount /var/lib/docker of that service to a specific location (e.g. the
build directory) with current Gitlab CI capabilities.

Running another container on the Gitlab DIND instance and bind mount
/var/lib/docker to gain access to the storage backend would be possible.
However, running another container on that Docker instance would pollute
the storage with an unnecessary (x86) container in /var/lib/docker. And
since we cannot easily untangle the data in the storage folder, it is not
possible to remove that container when creating the storage tarball.

The approach now taken is to run another DIND instance. This (DIND)IND
approach seems a bit excessive, but seems to work well. Since the Gitlab
DIND instance is running in a separate container we need to support a
"remote" Docker instance (by using the host part of DOCKER_HOST
environment variable). To simplify networking we just use "host"
networking (which is essentially networking of the Gitlab DIND
instance).

Luckily the build dir (e.g. /builds/stefan.agner/ostree-tools-containers/)
is shared between the Gitlab Job and Gitlab DIND container. This allows us
to store the final tarball in $(pwd)/output/ without further doing.

“There's a point where we just let the music take over everything.”

Related-to: TOR-37
Signed-off-by: Stefan Agner <stefan.agner@toradex.com>