torizoncore-builder: add custom Docker host workdir support
Use customer Docker host workdir support from dockerbundle module.
This allow to specify the current working directory location as seen
by the Docker host for the bundle subcommand. This is required to
use this script from within a container, which likely uses a different
location than the Docker daemon running outside of the container (e.g.
/workdir vs. /home/user/myworkdir).

Related-to: TOR-647
Signed-off-by: Stefan Agner <stefan.agner@toradex.com>