isolate: add missing files to ignore filelist
Add ipk-postinsts and fw_env.conf to ignore list. The latter is created
by a script run on first boot on Colibri iMX7. On all other modules it
should not be ignored, in thory. However, since probably no customer
need to customize this configuration file, excluding it is most likely
fine.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>