unpack/union/deploy: use archiv mode OSTree repository
Currently we operate in the bare mode repository part of the OSTree
sysroot in the unpacked Tezi image. This does not allow to pull the
OSTree via http, or upload the OSTree to our OTA servers using
garage-push. Start using a archive mode OSTree for all operations.

Note: We still use the sysroot to find the base OSTree hash to use
(from the deployment used in the original image). We also filter out the
kernel arguments (as they are only present in the deployment in OSTree)
as well as the original /home directories. Ideally we should get rid of
this dependency, but this need some rearchitecting.

Related-to: TOR-1237
Signed-off-by: Stefan Agner <stefan.agner@toradex.com>