union: use single ostree commit to overlay local changes
Instead of constructing a new OSTree from a complete checkout, just use
a single commit where we overlay the directory tree with the local
changes. This is a lot faster (0.02s vs. 8.5s).

The technique before has been used since it was recommended by the
OSTree documentation (see [0]).

However, the switch to a OSTree archive repository, and the
implementation of `--tree=ref` (see [1]) support for OSTree commit
changed trade-offs of the implementation variant significantly.
Today, there is almost no downside to use ostree commit directly.

[0] https://ostree.readthedocs.io/en/latest/manual/buildsystem-and-repos/
[1] https://github.com/ostreedev/ostree/commit/c7b12a8730260c66bc96363a3a3c5805b325bac8

Related-to: TOR-1238
Signed-off-by: Stefan Agner <stefan.agner@toradex.com>