Dockerfile: use requirements.txt
Create a requirements_debian.txt for Python packages we need on-top of
the ones we install from the Debian repository. It seems that setuptools
from the Debian repository is good enough, so use this instead of pip.
Also install wheel which fixes errors such as the following during pip
installation:
  error: invalid command 'bdist_wheel'

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>