.gitlab-ci.yml: use BuildKit for improved Docker cache usage
Currently the Docker build cache is not used due to a wrong variable
used as tag. Furthermore, because we use a multi-stage setup with an
intermediate stage which is not pushed to the registry, not all steps
can be used from cache (especially the rather costly building step).
And then there is also an intermediate stage which is not really used
to build the final image: tcbuilder-dev. Unfortunately, legacy docker
build is not smart enough to skip that build step.

Fix this issues by building and pushing the ostree-builder step
separately and use Docker's BuildKit which makes smarter decisions which
images to build.

Note 1: BuildKit needs --build-arg BUILDKIT_INLINE_CACHE=1, otherwise
caching information are not stored in the image, see:
https://github.com/docker/buildx#--cache-tonametypetypekeyvalue

Note 2: For some reason not-existing references in --cache-from seem
to be more problematic in BuildKit. However, I was not able to reproduce
this on my host. This might be due to an old Docker version on our build
servers (19.03.5 vs 19.03.12). Let's just use the build cache from the
local registry.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>