use f-string for logging
The attempt to fix Pylint warning "Use lazy % formatting in logging
functions [pylint(logging-fstring-interpolation)]" failed miserably:
Today the logging framework does not allow to use f-string style
string interpolation. It also would not work with the way attempted. If
anything **kwargs would need to be used to explicitly pass the name of
the strings to be interpolated. But this is also not supported by
th logging framework.

So the only options we have today is using % or %s style formatting.
Both seem not to be very appealing. So we decided to use f-string,
take the potential performance hit and disable the Pylint warning.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>