push: add push sub-command to push to OTA systems
The push sub-command uses garage-push and garage-sign to push a OSTree
to the OTA systems TreeHub server and signs the commit hash in
targets.json.

Unfortunately the Ubuntu packages provided by the upstream aktualizr
project do not install with Debian buster. So we have to build the
tools specifically for Debian. This seems to work quite flawless.

The sub-command by default uses the internal archive OSTree and reads
the commit metadata to find the hardware ID (typically the OE machine
name). However, both can optionally be provided as arguments. This
allows to use the sub-command to push the OpenEmbedded built OSTree as
well.

Related-to: TOR-1230
Signed-off-by: Stefan Agner <stefan.agner@toradex.com>