unpack/deploy: remove custom sysroot option
The custom sysroot option was mainly used during early development. It
defaults to sysroot under the storage directory (/storage/sysroot).
Today we also use other fixed paths underneath the storage directory,
e.g. tezi or ostree-archive. We would have to introduce custom arguments
for the too... But there is really not value in doing that.

Drop the custom sysroot option.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>