deploy: use checksum for deployment
The change to use "base" as reference actually lead to an error during
deployment:
  Commit has no requested ref ‘base’ in ref binding metadata

It seems that pull-local creates a new reference automatically, however,
since base is not in the ref-binding metadata it cannot create that
reference.

Ideally we should use the "real" reference (e.g. if there was no change
the original branch "5/colibri-imx6/torizon/torizon-core-docker/nightly"
or the user given branch name otherwise).

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>