deploy: implement deploy to device
Implement direct deployment to the device. This allows to quickly test
changes by actually deploying the OSTree to the device.

Use the "serve" backend to serve the OSTree archive repository via HTTP.
Use SSH port forwarding to make the HTTP server accessible from the
device. This avoids firewall or similar issues with the nice side effect
of having even an encrypted transport channel.

Use the same SSH connection to also execute ostree commands remotely to
fetch and deploy the OSTree.

Related-to: TOR-1284
Signed-off-by: Stefan Agner <stefan.agner@toradex.com>