torizoncore-builder: change default log level
Currently we set a single log level on root level, which is INFO. It
turns out that quite some libraries (e.g. aiohttp) are a bit noisy in
INFO level.

Let's default to WARNING as global setting and INFO for the TorizonCore
namespace.

Note: Just not touching the root level (as it was earlier) is not an
ideal situation as this will add a standard logger, which logs to
stderr (this then leads to a bit messy output). With the current
approach torizoncore-builder takes control over the full logging
behavior.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>