dockerbundle: improve logging in case tarball can't be found
We run a container to create the tarball and then expect it to be
available in the working directory of the script context. However, since
the script itself runs in a container and runs a container to execute
the tar command, it happens that the tar does not end up where we expect
it. Make debugging such issues easier by logging in case the tarball
does not appear where it should.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>