handle devicetree binary option for overlay more efficiently
We had implemented to present a list of available devicetree
binaries from OSTree and Volume to user, if no devicetree binary
is provided by user. This is not CI friendly, and won't work with
the torizoncore builder manifest file approach. Instead,if only
devicetree binary name is provided without any path, look inside
OSTree and if no devicetree is provided, pick default "devcietree"
binary file from OSTree. If user provides a devicetree binary from
working directory, look for it in working directory and if not
found, look inside OSTree.

Related-to: TOR-1435

Signed-off-by: Naeem Khan <naeem.khan@toradex.com>