union: implement whiteouts in Python
With the move from the checkout/union approach to the single commit
union approach we missed implementing a crucial feature: Whiteouts. It
seems that OSTree only supports whiteout in the checkout command.
However, thanks to OSTree.MutableTree it is rather trivial to implement
this functionality in Python.

This commit readds whiteout functionality. It processes whiteout files
after every changes directory. This means that the order of changes
directories applied matters, just like when using checkout or in
container world.

Related-to: TOR-1472
Signed-off-by: Stefan Agner <stefan.agner@toradex.com>