.gitlab-ci.yml: disable cached builds
The build of the container sometimes fails with the following error:

failed to compute cache key: failed to walk
/var/lib/docker/overlay2/df6ea.../merged/root/aktualizr/build: lstat
/var/lib/docker/overlay2/df6ea.../merged/root/aktualizr/build: no such file or directory

It seems a cache issue with the sota-builder build stage. For some
reason, using --cache-from with BUILDKIT_INLINE_CACHE does not work
every time.

Indeed, there is a report of this issue in [1] and it is fixed in [2],
but it seems our CI it's not running a Buildkit version that has this
fix.

To workaround this issue, let's disable cached builds for now.

[1] https://github.com/moby/buildkit/issues/1981
[2] https://github.com/moby/buildkit/pull/1993

Related-to: TOR-1671

Signed-off-by: Sergio Prado <sergio.prado@toradex.com>