tests/integration: introduce BATS framework
BATS (Bash Automated Testing System) is a TAP-compliant testing
framework for Bash. It provides a simple way to verify that the UNIX
programs you write behave as expected.

This commit introduces BATS as a framework to do integration tests in
TorizonCore Builder.

There is a script to configure the environment called setup.sh, a script
to download Easy Installer images called get_tezi_images.sh and a script
to run the tests called run.sh. Instructions on how to configure the
environment and run the test cases are available in the README file.

This commit introduces some test cases in testcases/. A lot of tests are
still missing, and will be introduced in future commits.

Lastly, this commit introduces a job in the CI script, so tests that
don't depend on a device will run for every commit in the repository.

Related-to: TOR-1565

Signed-off-by: Sergio Prado <sergio.prado@toradex.com>