bundle: Fix issue with bundle when running on Windows 10 WSL
By default "containers.run" returns the logs of the container upon container finish.
However for some reason when running on Windows there seems to be an issue where the container
can't be found because it exits and removes itself before the logs are fetched.
Therefore causing a "container can't be found" error.

Resolve this by adding the "detach=True" argument.
This causes "containers.run" to return a container object instead of the container logs.
Thus working around the error, since now the logs won't be fetched.

Related-to: TOR-1747

Signed-off-by: Jeremias Cordoba <jeremias.cordoba@toradex.com>