dto: Change handling of dto list when default device tree can't be determined
dto list will now list all found dtbs in the sysroot if a default can't be determined.
If only one is found then it will proceed using this as the assumed default.

If more than one is found they will be listed and the user can re-run
dto list and choose one with --device-tree, as the assumed default.

Related-to: TOR-1691

Signed-off-by: Jeremias Cordoba <jeremias.cordoba@toradex.com>