isolate: don't ignore users/groups files
A user may want to change users and groups directly in the device
and after that isolate the changes with TorizonCore Builder.

So this commit removes from the list of ignored files: gshadow, shadow
and group.

Also, let's split the list of ignored files one per line so it is
easier to do diffs in the future.

Related-to: TOR-1752

Signed-off-by: Sergio Prado <sergio.prado@toradex.com>