build: Validate configuration file against schema
In this commit the original code (by Valter Minute) for validating the
configuration file was reused (made part of the build backend) and the
original schema was adapted and augmented to match the new specification.
Also the code to handle variable assignments in the program command
line was moved to the backend.

Closes: TOR-1787

Related-to: TOR-1547

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>
Signed-off-by: Valter Minute <valter.minute@toradex.com>