images: No longer use work directory to store uncompressed files
With this commit the 'images unpack' command will no longer store the
extracted files from the compressed image into the work directory.
Instead those files will be stored in the storage directory (which is
usually kept in a Docker volume).

Although this is not really related to the configuration file
implementations of TOR-1547, I am implementing it in the same branch
because the the generation of those files causes some headaches when
you have both the compressed and the uncompressed files in the work
directory which is commonly the case when you are testing torizoncore-
builder with various types of inputs (as I am doing).

Closes: TOR-1796

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>