dt: Minor refactoring to simplify reuse
This is a simple refactoring to simplify the programmatic use of the 'dt'
command by the high-level 'build' command. I also took the opportunity
to clear the pylint warnings (most of them related to long lines). This
is part of the implementations for TOR-1808 (handle customization section
of the configuration file).

Related-to: TOR-1547

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>