build: Handle docker-compose file bundling
With this commit we finish handling the bundle subsection (belonging to
the output section). This means that the user can now specify a docker-
compose file (plus optional username, password and platform) and the
tool will automatically create the bundle and combine it with the final
image. To minimize code changes, the bundle is currently generated in
the work directory and removed after it has been combined with the image.

A further improvement: if something goes wrong while creating the output
image, the output is removed to avoid exposing the user to unfinished
artifacts.

Closes: TOR-1814

Related-to: TOR-1547

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>