build: Handle variable substitution and template generation
This commit introduces the variable substitution feature whereby shell-
like variable expansion notations such as $VAR, ${VAR} and similar
can be used within the configuration file. This is supposed to help
especially in CI scripts.

Another feature introduced is the command-line switch --create-template
that when passed by the user will trigger the generation of a template
configuration file.

Related-to: TOR-1547

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>