tests/integration: Update images unpack test case
Originally the 'images unpack' test case considered that the unpacked
image directory would be present in workdir after execution of the
unpack command on a tarred image. Since this does not happen anymore,
the test case was updated to first unpack the tarred image before
running the unpack command taking a directory as input.

Related-to: TOR-1547

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>