build: Determine image file name from Content-Disposition header
When the URL set in the 'remote' property in the input section of the
configuration file does not supply a name that looks like a tar or zip
file and the user has not specified a file name either, we try to
determine it from the Content-Disposition header that the server
usually provides. The implementation works with the legacy header only
(no support for internationalized names such as those supported by
RFC6266).

Closes: TOR-1840

Related-to: TOR-1547

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>