build: Provide better output messages upon parsing errors
With this commit we improve the program output when errors occur while
parsing the configuration file: now very large messages usually showing
parts of the schema are avoided; we also show which property caused the
error; when multiple errors are detected, we show all of them instead
of quitting at the first one as before.

Additionally, the --force flag that is taken by the build command is
being used also when the program is producing the template file; thus,
when the user passes --force --create-template the template will be
created even if it already exists.

Closes: TOR-1788

Related-to: TOR-1547

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>