build: Add integration tests and bug fix
Here we add integration tests for the "build" command along with a fix
for a bug found with one of the tests: when the configuration file had
errors in items belonging to a list the parsing errors would not be shown
(a stack trace was shown instead). This happened because we assumed that
the location of the error was always formed by strings whereas in this
case it contains a number representing the index of the failing element.

Closes: TOR-1547

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>