gitlab-ci: Log in to DockerHub to avoid pull limit errors
Because the integration tests of the bundle and combine commands are
accessing DockerHub, we were commonly getting pull limit errors during
those tests. To avoid this situation we are now logging in to
DockerHub. Additionally, the wget command line to fetch test images
was changed to print less dots during the download in order to take
up less space in the CI log.

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>