images: Get information about target from image-repo metadata
During tests with the offline-update endpoints we found that the custom
field that was supposed to be part of the offline-update metadata was
not filled out as originally expected by torizoncore-builder for
processing each target. The present commit tries to solve this by
looking up the target on the image-repo targets metadata (this
involves recursively going over the delegations).

Related-to: TOR-2032

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>