tests: Improve test of dt command
The test of the dt command that deploys an image to the device was
rewritten so that it no longer depends on the previous test. Also, now
it tries to detect the current device tree being used by the device and
then modifies that specific device tree; this solves a problem on the
Apalis iMX8 where the network interface of the device stopped working
due to the change of the device-tree being used.

Related-to: TOR-2136

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>