bundle: No longer depend on cgroups to determine container ID
With this commit, the bundle command should operate correctly under
cgroups v1 and v2; the method to determine the current container ID is
now independent of cgroups. Instead it tries to determine the ID based
on environment variable TCB_CONTAINER_NAME and it falls back to using
the well-known image name if that variable is not set.

Related-to: TCB-209

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>