build: Fix variable substitution feature
Originally the variable substitution operation was executed after the
validation against the schema which caused the validation to fail with
fields that were not free-form strings. Now, the validation is done
after variable substitution as it should be. This commit also adds the
tests to detect the original issue so as to prevent future regressions.

Related-to: TCB-251

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>