images: Implement provisioning command
Add new command "images provision" that allows a Toradex Easy Installer
image to be imbued with data for offline and/or online provisioning.
Besides this new command, new fields were added to the tcbuild file to
allow the feature to be employed with the "build" command. A series of
tests were also implemented to test the new command and its interaction
with relevant existing commands (combine and deploy).

The addition of this new feature required some rework of the function
responsible for adding files to the 'filelist' field in the image.json
file. The reason being that before there was an assumption that if the
'filelist' field was not empty then the image had containers in it
which is no longer true.

Related-to: TCB-248

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>