platform push: Handle canonicalization in more setups
Basically the canonicalization code was rewritten to handle the various
cases that we were expected to support, namely:

- Default registry with and without authentication;
- Known registries (those whose certificate is verifiable by the CAs
  bundled with the distro, e.g. GCR, GitLab, etc.) with and without
  authentication;
- Secure company-hosted registries (rather, those whose certificate
  for verification must be passed to TorizonCore Buider) with and
  without authentication;
- Insecure registries (accessed via HTTP);

For the insecure case, we follow a "fall-back" approach where we try
to access the registry via HTTPS and then switch to HTTP in case of
SSL errors; this is to avoid the need for the users to pass an extra
parameter to the tool in order to inform it about the insecure
registries. This could be seen as a security flaw but since an
attacker would not be able to install the images on the device later
we considered this an acceptable risk.

Related-to: TCB-371

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>