registryops: Handle repo names with a dot in them in platform lockbox
Pass full repository name (including registry and namespace name) to
the manifest fetching call-chain employed by "platform lockbox"; this
is to avoid problems with names having a dot in them (particularly in
their namespace name).

For example, a name like "gitlab.com:5000/some.namespace/someimage" did
not work before due to the dot in "some.namespace" and with this change
it does.

Related-to: TCB-381

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>