tests: Changes to avoid DockerHub pull limit erros
We are getting pull limit erros during the integration tests. The issues
are related to one of the docker-compose used in the build command; and
the images pulled from DockerHub and used to create the private registry
containers.

A new docker-compose file has been added to prevent the build test from
fetching images form DockerHub when it is not possible to login to it.

A docker login command has been added to the DinD container that populates
the private registry containers.

Signed-off-by: Nathan Matos de Almeida <nathan.matos@toradex.com>