tests: Imporve platform push tests
When pushing files with the 'platform push' command with the same
package name, uptane-sign will fail with an overwrite error if a
previous version name was already used. To avoid this, we used to
generate the version name based on the current time only. Imporving
the version name generation to also add the date.

To also avoid conflicts with package name and version, they will start
with the value from the MACHINE environment variable from the run.sh
script.

Jumping some 'platform push' tests due to the possibility of a racing
conditions among the pipelines if using the same set of credentials.

Signed-off-by: Nathan Matos de Almeida <nathan.matos@toradex.com>