Dockerfile: No longer force Git version
To avoid failing on a "dt checkout" update (i.e. when the --update
switch is specified) we now configure Git to consider the device-trees
directory as "safe". This way we no longer need to force the use of an
older Git version which does not do this safety check. That is good
because the version we specified was no longer available which caused
TorizonCore Builder builds to break.

Related-to: TCB-373

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>