platform push: Implement generic package push
Generic packages are defined as all the packages that have
hardwareids other than the primary and secondary docker-compose.

The system to identify a generic package uses a very simple
heuristic technique. The REF command line parameter, if it is not
a file is considered an ostree reference so the system performs a
push_ref, on the other hand, if the REF parameter is a file and ends
with .yml or .yaml it is considered a docker-compose package,
otherwise a file with any other extension is considered a
generic package.

For the generic package, the parameter --hardwareid is mandatory.
This commit also implements a new option --custom-meta that is used
by the push_generic() to add JSON metadata to the director.

This commit adds also a preliminary implementation of push_generic
integration tests.

Related-to: TCB-356

Signed-off-by Stefano Viola <stefano.viola@toradex.com>