tests/integrations: Fix bug in function.sh
The torizoncore-bulder() function, uses $@ without quotes
to read all function parameters, but this is not the best
way to handle function prameters.
If a parameter contains, for example, spaces, this will surely generate
unexpected behavior.

Using "$@" instead of $@ is considered a best practice to handle arguments
correctly, especially when dealing with parameters that may contain spaces,
special characters. It helps ensure that the arguments are passed to
subsequent commands or functions as intended, without unintended side effects.

Because of this change, there was a need to fix other tests.

Signed-off-by Stefano Viola <stefano.viola@toradex.com>