platform push: Implement compatible-with for generic packages
This commit implements the parameter --compatible-with for the push_generic()
This allow the user to restrict a generic package to a particular OS image.
The specified OS image must be accessible in the Platform.

This commit also implements additional tests for the push_generic():
- Test with compatible-with parameter
- Test with compatible-with and custom-metadata parameter
- Test with some invalid custom metadata.

Related-to: TCB-356

Signed-off-by Stefano Viola <stefano.viola@toradex.com>