bats: Replace 'load' syntax with 'bats_load_library' for relative paths
In Bats version v1.6.0, the behavior of the 'load' syntax was modified.
Previously, when given a relative path, the 'load' syntax searched for
libraries in the directory where the command was executed.

This commit introduces the use of 'bats_load_library' to handle relative paths.
When utilizing a relative path, 'bats_load_library' now relies on the
'BATS_LIB_PATH' variable to determine the library location.

Related-to: TCB-400

Signed-off-by: Nathan Matos de Almeida <nathan.matos@toradex.com>