images: Support symlinks to Tezi tarballs
The filename of the tarball is used to predict the name of the directory
contained in the tarball. When a symlink is used with a different
(simpler) name, the unpack would fail.

This resolves the symlink and looks at the filename of the actual
tarball file.

This can be useful to allow pointing tcbuild.yaml to a symlink with
a stable name, which points to a locally built tarball which changes
name on every build. For example, see this patch that does this
automatically for the yocto builds:

    https://community.toradex.com/t/patch-create-non-versioned-symlink-to-tezi-tarball/21265

Signed-off-by: Matthijs Kooijman <matthijs@stdin.nl>