Use tar --use-compress-program option and run without shell
This removes the need for building a decompression pipeline, but lets
tar handle this internally. This in turn removes the need to run these
commands through the shell and allows passing the commands to run as
arrays without relying on shell word splitting and other
interpretations.

In practice, this prevents special characters such as spaces from
breaking the build.

This was tested with all four code paths and gzip, xz, bzip2 and no
compression (but not all combinations).

Signed-off-by: Matthijs Kooijman <matthijs@stdin.nl>