kernel: Do not print backtrace on failed module build
This backtrace does not really add much except noise, and it is already
printed at DEBUG level higher up.

The exception message, which includes the failing command and exit code,
is still useful, so include that in the error message.

This also limits the except clause to CalledProcessError, since that is
the only exception we should be expecting here and thus the only one to
replace with a TorizonCoreBuilderError. If any other exception occurs,
it is just propagated and handled by the top-level script (with
a backtrace).

Signed-off-by: Matthijs Kooijman <matthijs@stdin.nl>