Run subprocesses without shell where possible
In a lot of places, subprocesses were spawned with shell=True causing
an extra shell invocation to interpret a string commandline. This means
that any shell special characters and spaces in the command will be
interpreted by the shell. Since these were usually not escaped, using
shell special characters or spaces in filenames would cause commands to
fail.

This commit changes many of these invocations to not use the shell, but
instead specify the command to execute as an array of arguments. This
means no additional word-splitting is needed, so special characters and
spaces work as intended.

There are still some subprocess calls that still use the shell because
they spawn a pipeline of multiple commands. These will be handled in the
next commits.

Most of these code paths were tested, except some of the dt/dto ones.

Signed-off-by: Matthijs Kooijman <matthijs@stdin.nl>