Quote variables used in shell invocations
For the remaining shell invocations, this makes sure that any variables
used inside the shell commandlines are quoted using shlex.quote. This
should ensure that any variables (filenames) containing spaces or other
special characters are appropriately escaped for use with
a POSIX-compatible shell.

Note that these invocations cannot be easily convert to not use a shell
at all, since they spawn a pipeline consisting of two or more piped
commands. It is possibe to build an equivalent pipeline from within
python without needing the shell, but that is left as a future
improvement. See also:

    https://docs.python.org/3/library/subprocess.html#replacing-shell-pipeline

Signed-off-by: Matthijs Kooijman <matthijs@stdin.nl>