backend/union: Do not raise error if file tagged for removal isn't found
Currently if TCB tries to delete a file in changes_dir that doesn't
exist in the unpacked image, it raises an exception and the program is
completely halted.

This commit changes this behavior so that TCB will only print a warning
message when the above situation happens, ignoring the non-existent file
and continuing execution as expected.

Signed-off-by: Lucas Akira Morishita <lucas.morishita@toradex.com>