build: Add initial support for WIC/raw images
It is now possible to use the build command to unpack, modify and
locally deploy a Torizon OS image in .wic/.img format.

Modification of raw images are still limited to the currently supported
standalone commands. Currently only splash screen and /etc modification
are confirmed to work, with the rest of the other features coming later.

This commit adds a new 'raw-image' field in the input and output
sections of tcbuild.yaml, where the user can put all the info needed to
build a custom image with this format.

There are checks that verify if the image type described in the input
section has a corresponding entry in the output section.

The tcbuild.yaml template file was also updated to have the new options
and details on how each section works.

Signed-off-by: Lucas Akira Morishita <lucas.morishita@toradex.com>