kernel: Remove dependency on value of CONFIG_LOCALVERSION
Currently TCB uses the value of LOCALVERSION to get the major version of
the unpacked Torizon OS image. With it TCB determines the toolchain used
to build kernel modules.

The image major version is also used to decide whether or not to apply a
specific patch in the kernel source Makefile.

In both cases we don't need LOCALVERSION:

- We can get the major version of the unpacked image more reliably by
  using an already existing function in common.py called
  get_branch_and_major_from_metadata().

- The Makefile patch is applied due to a change introduced in kernel
  6.1. Therefore we need to check the kernel version, not the image
  version.

This commit replaces all usage of LOCALVERSION by using the
alternatives described above.

Signed-off-by: Lucas Akira Morishita <lucas.morishita@toradex.com>