Show spinner loading animation when processing raw images
Show a spinning wheel animation on the terminal during some
time-consuming tasks to indicate that TCB is actively running something
and not stuck. This implementation runs a given function while the
spinning animation is being executed at the same time in a different
thread.

This should provide users with a visual feedback that the program is
working as it should, and not frozen.

Signed-off-by: Lucas Akira Morishita <lucas.morishita@toradex.com>