Hotfix for Python lib requests 2.32.0
requirements_debian.txt: Pin 'docker' version to 7.1.0 and pin
'requests' version to 2.32.2.

Currently when building the TCB container, the latest version of
the 'requests' Python library is downloaded as part of the dependency
chain of the packages listed in requirements_debian.txt.

The recently released requests 2.32.0 breaks Docker SDK for Python,
which causes errors when doing a bundle command. More info about the
requests 2.32.0 issue can be found here:

https://github.com/docker/docker-py/issues/3256

This was fixed in 2.32.2, but to use it docker-py also needs to be
updated to at least version 7.1.0.

Signed-off-by: Lucas Akira Morishita <lucas.morishita@toradex.com>