CI tests: Add support for 'torizon-docker' images
Starting with Torizon OS 7, the default Yocto image name was changed
from 'torizon-core-docker' to 'torizon-docker'. .gitlab.yml and
pr-tests.yml are not able to correctly identify the machine with
this new naming format.

And in particular, when downloading images from Artifactory, the script
assumes that 'torizon-core-docker' is always part of the download link,
which is no longer the case.

This commits aims to fix the above while maintaining compatibility with
Torizon OS 6 images.

Signed-off-by: Lucas Akira Morishita <lucas.morishita@toradex.com>