kernel: Restrict kernel Makefile hotfix for 6.1 only
In order to build kernel modules, TCB applies a small change to the
Makefile for kernel version 6.1, used in Torizon OS 6 images. Without it
the modules_prepare task of the Makefile gets stuck into a recursion
loop.

It was assumed that this behavior would also apply for versions above
6.1. However, this patch seems to cause problems with kernel 6.6 used in
Torizon OS 7 images. In fact, 6.6 works without any modifications to the
Makefile.

It is suspected that this recursion loop happened in the first place due
to how the kernel sources are bundled with the image i.e. maybe some
Kbuild/Makefiles are missing.

This commit restricts the Makefile patch to only be applied for kernel
version 6.1. Other versions are left unaltered.

Signed-off-by: Lucas Akira Morishita <lucas.morishita@toradex.com>