bundle: Remove dependency on compose V1 library
Currently the bundle command directly uses the deprecated docker compose
V1 Python library to parse data of the input docker compose YAML file.

Although the bundle command basically only needs to know the container
image names in the compose file in order to download them, the library
calls end up checking if the whole file follows the compose V1
specification.

Because of this, files with specific compose V2 features will fail the
above check, and the bundle command will be aborted.

This commit removes the usage of this library, and apart from a basic
check of container image names, TCB will no longer attempt to validate
the entire compose file.

As a side effect of not using the compose V1 library, environment
variables referenced in the compose YAML file will no longer be
expanded.

Signed-off-by: Lucas Akira Morishita <lucas.morishita@toradex.com>